requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
bpemb==0.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.1
Deprecated==1.2.7
docutils==0.15.2
flair==0.4.4
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
gast==0.2.2
gensim==3.8.1
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
hyperopt==0.2.3
idna==2.8
importlib-metadata==1.4.0
ipython==7.6.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rcnn==0.0.2
keras-resnet==0.2.0
keras-rl==0.4.2
keras-tqdm==2.0.1
kiwisolver==1.1.0
langdetect==1.0.7
lazy-load==0.8.2
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
mpld3==0.3
networkx==2.2
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.1
pandas==0.25.3
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.13.1
prompt-toolkit==2.0.10
protobuf==3.11.2
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pymongo==3.10.1
PyMuPDF==1.16.10
pyparsing==2.4.6
# Editable install with no version control (pysoul==0.1)
# -e /home/vinaycr7/Desktop/ge
pytest==5.3.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.2
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
segtok==1.5.7
sentencepiece==0.1.85
six==1.14.0
sklearn==0.0
smart-open==1.9.0
sqlitedict==1.6.0
tabulate==0.8.6
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tiny-tokenizer==3.1.0
torch==1.3.0
torchvision==0.4.1
tqdm==4.42.0
traitlets==4.3.3
transformers==2.3.0
urllib3==1.24.3
wcwidth==0.1.8
Werkzeug==0.16.1
wrapt==1.11.2
zipp==2.1.0
